- Separate components - Add update of draft - List of drafts
