+++ packages/pdf-api/src/draft/draft.controller.ts
@@ -16,7 +16,7 @@ const logger = getLogger('DraftController');
 @Controller('api/')
 @RequireApiKey()
 export class DraftController {
-  constructor(private readonly draftServie: DraftService) {}
+  constructor(private readonly draftService: DraftService) {}
 
   @Put('users/:userName/drafts/:draftName')
   @HttpCode(200)
@@ -60,7 +60,7 @@ export class DraftController {
   ) {
     logger.debug('saveDraft', { draftName, userName });
 
-    await this.draftServie.saveDraft({
+    await this.draftService.saveDraft({
       userName,
       draftName,
       params: draftParams,
@@ -74,7 +74,7 @@ export class DraftController {
   ): Promise<DraftDetails> {
     logger.debug('getDraft', { draftName, userName });
 
-    const draftDetails = await this.draftServie.getDraft({
+    const draftDetails = await this.draftService.getDraft({
       userName,
       draftName,
     });
